Moved cg and gmres matrix market tests to public API
